Refactoring TabsArea and TabsAreaLayout to a different directory. #54
